Set domains to empty array when not in config file.
In order to simplify the logic, this proposed change simply creates an empty domains array if none are found in the config file. This way, if we check the length of our "whitelist" and there are none, we assume we don't want to continue checking the domain to match at all (emails of all domains are allowed).